// src/Model/iTechRestApi/iTechDataTaskEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechDataTaskEnum {
    root = 1,
    groupHeader = 2,
    objectReviewManual = 3,
    verificationAcknowledgement = 4,
    assigningSupervisor = 5,
    idVerification = 6,
    dataDiscoveryManual = 7,
    fullyAutoDeletion = 8,
    notifySarRequestor = 9,
    notifySupervisor = 10,
    closeWorkflow = 11,
    assignTeam = 12,
    supervisorReview = 13,
    requestorResponse = 14,
    amendSystem = 15,
    requestID = 16,
    idOutcome = 17,
    idSuccessInitiation = 18,
    discoveryRequest = 19,
    discoveryEmailRequest = 20,
    discoveryRequestResponse = 21,
    objectReviewAutomatic = 22,
    workflowParameters = 23,
    susActivityVerification = 24,
    processorErasureNotification = 25,
    notifyParties = 26,
    dataDiscoveryAutomatic = 27,
    dataSourceSelection = 28,
    dataDiscoveryHeader = 29,
    dataReviewHeader = 30,
    dataSourceReviewHeader = 31,
    selectReviewPercentage = 32,
    taskAssignmentAutomatic = 33,
    termSelection = 34,
    termAddition = 35,
    redactionExecution = 36,
    sendOutput = 37,
    rectificationExecution = 38,
    agentReviewHeader = 39,
    agentApproval = 40,
    fileUpload = 41,
    alertReview = 42,
    form = 43,
    closeCaseManual = 44,
    alertReviewL2 = 45
}