// src/Model/iTechRestApi/iTechDataCaseEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechDataCaseEnum {
    default = 1,
    surveillance = 2,
    gdpr = 3,
    behaviour = 4,
    trade = 5,
    hr = 6,
    financialCrime = 7,
    conduct = 8,
    business = 9,
    internal = 10
}