// src/Model/iTechRestApi/iTechDataCaseSubEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechDataCaseSubEnum {
    default = 0,
    art15_Access = 1,
    art16_Rectification = 2,
    art17_Erasure = 3,
    art18_RestrictProcess = 4,
    art19_Notification = 5,
    art20_RightToDataPortability = 6,
    art21_RightToObject = 7,
    art22_RightToDataFromAutoDec = 8,
    defaultSurveillance = 9,
    suspiciousTradeActivity = 10,
    financialCrime = 11,
    businessConduct = 12,
    interestConflict = 13,
    customerComplaint = 14,
    employeeConduct = 15,
    hr = 16,
    infoBarrier = 17,
    infoSecurity = 18,
    marketConduct = 19,
    nonPublicInfo = 20,
    other = 21,
    riskInsight = 22,
    salesPractice = 23,
    frontRunning = 24,
    layering = 25,
    spoofing = 26,
    washTrading = 27,
    parking = 28,
    insiderDealing = 29,
    largeOrder = 30,
    markingOpen = 31,
    markingClose = 32,
    markingUpDown = 33,
    offMarketPricing = 34,
    momentumIgnition = 35,
    paintingTape = 36,
    exfiltration = 37,
    pnl = 38,
    toxicity = 39,
    fileAccess = 40,
    absences = 41,
    missedTraining = 42,
    doorEntrySwipes = 43,
    logInfo = 44,
    telemetry = 45,
    dpia = 46,
    art15_Access_Auto = 47
}